Add files via upload
Laboratory activity was conducted face-to-face.
This activity is about 1 group was assigned in building code in Arduino for Button, the group will connect to the Wi-Fi Connection of another iot device that has Arduino and code for Assigning LED pins. This activity objectives are to build connection between 2 device that is connected to 1 Wi-Fi or hotspot connection.